app/requirements.txt
PyJWT==2.6.0
boto3==1.26.6